+++ api/typescript/src/engine/kurtosis_engine_rpc_api_bindings/engine_service_pb.d.ts
@@ -1,6 +1,7 @@
 import * as jspb from 'google-protobuf'
 
 import * as google_protobuf_empty_pb from 'google-protobuf/google/protobuf/empty_pb';
+import * as google_protobuf_timestamp_pb from 'google-protobuf/google/protobuf/timestamp_pb';
 
 
 export class GetEngineInfoResponse extends jspb.Message {
@@ -147,6 +148,11 @@ export class EnclaveInfo extends jspb.Message {
   hasApiContainerHostMachineInfo(): boolean;
   clearApiContainerHostMachineInfo(): EnclaveInfo;
 
+  getCreationTime(): google_protobuf_timestamp_pb.Timestamp | undefined;
+  setCreationTime(value?: google_protobuf_timestamp_pb.Timestamp): EnclaveInfo;
+  hasCreationTime(): boolean;
+  clearCreationTime(): EnclaveInfo;
+
   serializeBinary(): Uint8Array;
   toObject(includeInstance?: boolean): EnclaveInfo.AsObject;
   static toObject(includeInstance: boolean, msg: EnclaveInfo): EnclaveInfo.AsObject;
@@ -162,6 +168,7 @@ export namespace EnclaveInfo {
     apiContainerStatus: EnclaveAPIContainerStatus,
     apiContainerInfo?: EnclaveAPIContainerInfo.AsObject,
     apiContainerHostMachineInfo?: EnclaveAPIContainerHostMachineInfo.AsObject,
+    creationTime?: google_protobuf_timestamp_pb.Timestamp.AsObject,
   }
 }
 
@@ -255,25 +262,25 @@ export namespace CleanResponse {
   }
 }
 
-export class GetUserServiceLogsArgs extends jspb.Message {
+export class GetServiceLogsArgs extends jspb.Message {
   getEnclaveId(): string;
-  setEnclaveId(value: string): GetUserServiceLogsArgs;
+  setEnclaveId(value: string): GetServiceLogsArgs;
 
   getServiceGuidSetMap(): jspb.Map<string, boolean>;
-  clearServiceGuidSetMap(): GetUserServiceLogsArgs;
+  clearServiceGuidSetMap(): GetServiceLogsArgs;
 
   getFollowLogs(): boolean;
-  setFollowLogs(value: boolean): GetUserServiceLogsArgs;
+  setFollowLogs(value: boolean): GetServiceLogsArgs;
 
   serializeBinary(): Uint8Array;
-  toObject(includeInstance?: boolean): GetUserServiceLogsArgs.AsObject;
-  static toObject(includeInstance: boolean, msg: GetUserServiceLogsArgs): GetUserServiceLogsArgs.AsObject;
-  static serializeBinaryToWriter(message: GetUserServiceLogsArgs, writer: jspb.BinaryWriter): void;
-  static deserializeBinary(bytes: Uint8Array): GetUserServiceLogsArgs;
-  static deserializeBinaryFromReader(message: GetUserServiceLogsArgs, reader: jspb.BinaryReader): GetUserServiceLogsArgs;
+  toObject(includeInstance?: boolean): GetServiceLogsArgs.AsObject;
+  static toObject(includeInstance: boolean, msg: GetServiceLogsArgs): GetServiceLogsArgs.AsObject;
+  static serializeBinaryToWriter(message: GetServiceLogsArgs, writer: jspb.BinaryWriter): void;
+  static deserializeBinary(bytes: Uint8Array): GetServiceLogsArgs;
+  static deserializeBinaryFromReader(message: GetServiceLogsArgs, reader: jspb.BinaryReader): GetServiceLogsArgs;
 }
 
-export namespace GetUserServiceLogsArgs {
+export namespace GetServiceLogsArgs {
   export type AsObject = {
     enclaveId: string,
     serviceGuidSetMap: Array<[string, boolean]>,
@@ -281,21 +288,25 @@ export namespace GetUserServiceLogsArgs {
   }
 }
 
-export class GetUserServiceLogsResponse extends jspb.Message {
-  getUserServiceLogsByUserServiceGuidMap(): jspb.Map<string, LogLine>;
-  clearUserServiceLogsByUserServiceGuidMap(): GetUserServiceLogsResponse;
+export class GetServiceLogsResponse extends jspb.Message {
+  getServiceLogsByServiceGuidMap(): jspb.Map<string, LogLine>;
+  clearServiceLogsByServiceGuidMap(): GetServiceLogsResponse;
+
+  getNotFoundServiceGuidSetMap(): jspb.Map<string, boolean>;
+  clearNotFoundServiceGuidSetMap(): GetServiceLogsResponse;
 
   serializeBinary(): Uint8Array;
-  toObject(includeInstance?: boolean): GetUserServiceLogsResponse.AsObject;
-  static toObject(includeInstance: boolean, msg: GetUserServiceLogsResponse): GetUserServiceLogsResponse.AsObject;
-  static serializeBinaryToWriter(message: GetUserServiceLogsResponse, writer: jspb.BinaryWriter): void;
-  static deserializeBinary(bytes: Uint8Array): GetUserServiceLogsResponse;
-  static deserializeBinaryFromReader(message: GetUserServiceLogsResponse, reader: jspb.BinaryReader): GetUserServiceLogsResponse;
+  toObject(includeInstance?: boolean): GetServiceLogsResponse.AsObject;
+  static toObject(includeInstance: boolean, msg: GetServiceLogsResponse): GetServiceLogsResponse.AsObject;
+  static serializeBinaryToWriter(message: GetServiceLogsResponse, writer: jspb.BinaryWriter): void;
+  static deserializeBinary(bytes: Uint8Array): GetServiceLogsResponse;
+  static deserializeBinaryFromReader(message: GetServiceLogsResponse, reader: jspb.BinaryReader): GetServiceLogsResponse;
 }
 
-export namespace GetUserServiceLogsResponse {
+export namespace GetServiceLogsResponse {
   export type AsObject = {
-    userServiceLogsByUserServiceGuidMap: Array<[string, LogLine.AsObject]>,
+    serviceLogsByServiceGuidMap: Array<[string, LogLine.AsObject]>,
+    notFoundServiceGuidSetMap: Array<[string, boolean]>,
   }
 }