+++ api/typescript/src/engine/lib/kurtosis_context/generic_engine_client.ts
@@ -8,8 +8,7 @@ import {
     GetEnclavesResponse,
     GetEngineInfoResponse,
     StopEnclaveArgs,
-    GetUserServiceLogsArgs,
-    GetUserServiceLogsResponse,
+    GetServiceLogsArgs,
 } from "../../kurtosis_engine_rpc_api_bindings/engine_service_pb";
 import {Readable} from "stream";
 
@@ -20,5 +19,5 @@ export interface GenericEngineClient {
     stopEnclave(stopEnclaveArgs: StopEnclaveArgs): Promise<Result<null, Error>>
     destroyEnclave(destroyEnclaveArgs: DestroyEnclaveArgs): Promise<Result<null, Error>>
     clean(cleanArgs: CleanArgs): Promise<Result<CleanResponse, Error>>
-    getUserServiceLogs(getUserServiceLogsArgs: GetUserServiceLogsArgs): Promise<Result<Readable, Error>>
+    getServiceLogs(getServiceLogsArgs: GetServiceLogsArgs): Promise<Result<Readable, Error>>
 }