+++ docs/changelog.md
@@ -1,14 +1,29 @@
 # TBD
 
+# 0.52.5
+
+### Fixes
+- Renamed `files_artifact_mount_dirpaths` to just `files`
+
+# 0.52.4
+### Features
+- Added the enclave's creation time info which can be obtained through the `enclave ls` and the `enclave inspect` commands
+
+### Fixes
+- Smoothened the experience `used_ports` -> `ports`, `container_image_name` -> `name`, `service_config` -> `config`
+
+# 0.52.3
+
 ### Changes
 - Cleanup Startosis interpreter predeclared
 
 # 0.52.2
 
 # 0.52.1
-
 ### Features
 - Add `wait` and `define` command in Startosis
+- Added `not found service GUIDs information` in `KurtosisContext.GetServiceLogs` method
+- Added a warning message in `service logs` CLI command when the request service GUID is not found in the logs database
 - Added ip address replacement in the JSON for `render_template` instruction
 - Implemented a `ToTargetArtifactUuid` version of a few service_network functions
 - Made `render_templates`, `upload_files`, `store_Files_from_service` accept `artifact_uuid` and
@@ -52,7 +67,7 @@ return `artifact_uuid` during interpretation time
 # 0.51.11
 ### Features
 - Improve how kurtosis instructions are canonicalized with a universal canonicalizer. Each instruction is now printed on multiple lines with a comment pointing the to position in the source code.
-- Support `private_ip_address_placeholder` to be passed in `service_config` for `add_service` in Starlark
+- Support `private_ip_address_placeholder` to be passed in `config` for `add_service` in Starlark
 
 ### Changes
 - Updated how we generate the canonical string for Kurtosis `upload_files` instruction
@@ -134,7 +149,7 @@ return `artifact_uuid` during interpretation time
 # 0.51.1
 - Added an `exec` command to Startosis
 - Added a `store_files_from_service` command to Startosis
-- Added the ability to pass `files_artifact_mount_dirpaths` to the service config
+- Added the ability to pass `files` to the service config
 - Added a `read_file` command to Startosis
 - Added the ability to execute local modules in Startosis